requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Login
Flask-Mail
Flask-Security-Too
Flask-Admin
Pillow
pytesseract
Flask-Reuploaded
openpyxl
Flask-WeasyPrint
Flask-SocketIO
eventlet
Bootstrap-Flask
Flask-Talisman
Flask-Limiter
python-dotenv
Flask-Caching
Flask-Babel
pytest
pytest-flask
PyMySQL